Elements implementing a dunder repr method
For python interface users it is typically convenient to see
the properties of the object if printed.
Added a repr method to the ElemType and derived classes. This
will now show the information similar to the .print() method
automatically if an element is printed.

C++ part could make use of that implementiong ostream operator.